requirements.txt
certifi==2020.11.8
chardet==3.0.4
idna==2.10
poloniexapi==0.5.7
requests==2.24.0
six==1.15.0
urllib3==1.25.11
websocket-client==0.57.0
